% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_siblings}
\alias{add_siblings}
\title{Create multiple siblings for an Individual}
\usage{
add_siblings(tg, xref, sexes = NULL)
}
\arguments{
\item{tg}{A tidyged object.}

\item{xref}{The xref of an Individual record to add siblings for.}

\item{sexes}{A character string giving the sexes of each sibling. For example,
"FFM" to add two sisters and one brother. See the help for tidyged::add_indi()
for possible codes.}
}
\value{
A tidyged object with additional sibling records.
}
\description{
Create multiple siblings for an Individual
}
