% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sense_checks.R
\name{age_now}
\alias{age_now}
\title{Determine the age of an individual now given their age on a previous date}
\usage{
age_now(date_of_fact, age_at_fact, minimise = TRUE)
}
\arguments{
\item{date_of_fact}{A date string from the tidyged object.}

\item{age_at_fact}{An age at event string from the tidyged object.}

\item{minimise}{If date ranges or periods are used in the date, whether to choose the bounds which
assume the minimum age. If this is FALSE, the maximum age is assumed.}
}
\value{
A numeric value giving the current age in years. A numeric value less than zero means no
determination could be made.
}
\description{
Determine the age of an individual now given their age on a previous date
}
