% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arrange_records}
\alias{arrange_records}
\title{Arrange all records in a tidyged object}
\usage{
arrange_records(gedcom, order = "IFMSRN")
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{order}{A character string indicating the desired order of records. The letters
indicate (I)ndividual, (F)amily Group, (M)ultimedia, (S)ource, (R)epository, (N)ote.}
}
\value{
An arranged tidyged object.
}
\description{
Arrange all records in a tidyged object
}
