% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sense_checks.R
\name{date_diff}
\alias{date_diff}
\title{Determine the number of years between two dates}
\usage{
date_diff(date1, date2 = tidyged::date_current(), minimise = TRUE)
}
\arguments{
\item{date1}{A date string from the tidyged object.}

\item{date2}{A date string from the tidyged object. If no date is given, today's date is used.}

\item{minimise}{If date ranges or periods are used in the dates, whether to choose the bounds which
assume the minimum date difference. If this is FALSE, the maximum date difference is assumed.}
}
\value{
A numeric value giving the number of years. A numeric value less than zero means no
determination could be made.
}
\description{
Determine the number of years between two dates
}
\details{
Doesn't yet handle dual years or BCE dates.
}
