% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_age.R
\name{guess_age}
\alias{guess_age}
\title{Guess an individual's age}
\usage{
guess_age(tg, xref, agg_fn = mean)
}
\arguments{
\item{tg}{A tidyged object.}

\item{xref}{The xref of an individual.}

\item{agg_fn}{If multiple ages are calculated, the function to use to aggregate them.}
}
\value{
A numeric value giving the estimated age. A numeric value less than zero means no
determination could be made.
}
\description{
This function calculates an age for an individual based on their individual or family facts. It calculates age based on the date of the fact and their age when the fact applied.
}
