% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_explicit_marr_types_all}
\alias{insert_explicit_marr_types_all}
\title{Insert explicit marriage subrecords}
\usage{
insert_explicit_marr_types_all(tg)
}
\arguments{
\item{tg}{A tidyged object.}
}
\value{
An updated tidyged object with additional marriage subrecords.
}
\description{
This function inserts explicit marriage subrecords for Family Group records that do not have one.
}
\details{
The GEDCOM specification recommends that all marriage events have an explicit TYPE subrecord
subordinate to the MARR relationship event. If one is not given, marriage is assumed.
}
