% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_alive}
\alias{is_alive}
\title{Determine whether a person is still alive from their date of birth}
\usage{
is_alive(date_of_birth, max_age)
}
\arguments{
\item{date_of_birth}{A date string from the tidyged object.}

\item{max_age}{The maximum age to assume for a living person.}
}
\value{
A logical value indicating whether the person is still alive.
}
\description{
Determine whether a person is still alive from their date of birth
}
