% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gedcoms.R
\name{make_xrefs_unique}
\alias{make_xrefs_unique}
\title{Update xrefs in a tidyged object to make them unique from another tidyged object}
\usage{
make_xrefs_unique(tg1, tg2)
}
\arguments{
\item{tg1}{The first tidyged object.}

\item{tg2}{The second tidyged object. This will be the one that will be updated.}
}
\value{
The second tidyged object, where all xrefs have been made unique from the first object.
}
\description{
This function ensures that all record xrefs across two tidyged objects are unique.
}
