% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gedcoms.R
\name{merge_records}
\alias{merge_records}
\title{Combine multiple records into a single record}
\usage{
merge_records(gedcom, xrefs)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xrefs}{The xrefs of the records to merge.}
}
\value{
A new tidyged object where all specified records have been merged.
}
\description{
This function takes multiple records and replaces them with a single record containing
all of their subrecords. It does not remove duplicate subrecords.
}
