% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gedcoms.R
\name{migrate_records}
\alias{migrate_records}
\title{Copy all records from one tidyged object to another}
\usage{
migrate_records(tg1, tg2)
}
\arguments{
\item{tg1}{The first tidyged object and the one you want to use for header and submitter information.}

\item{tg2}{The second tidyged object.}
}
\value{
A tidyged object which contains the same header and submitter information as the first object
and all records contained within both input objects.
}
\description{
This function takes two tidyged objects and copies all of the records in the second to the first,
while making their identifiers unique (even if they may be duplicates).
}
