% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_famg_children_all}
\alias{order_famg_children_all}
\title{Order children in all Family Group records by birth date}
\usage{
order_famg_children_all(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
The same tidyged object with rearranged children rows in the Family Group records.
}
\description{
Order children in all Family Group records by birth date
}
