% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gedcoms.R
\name{potential_duplicates}
\alias{potential_duplicates}
\title{Identify potentially duplicate records}
\usage{
potential_duplicates(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
The same tidyged object, potentially with some records merged.
}
\description{
Identify potentially duplicate records
}
