% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_change_dates}
\alias{remove_change_dates}
\title{Remove all CHANge dates from a tidyged object}
\usage{
remove_change_dates(tg)
}
\arguments{
\item{tg}{A tidyged object.}
}
\value{
A tidyged object with all CHAN structures removed.
}
\description{
Remove all CHANge dates from a tidyged object
}
