% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gedcoms.R
\name{remove_duplicate_subrecords}
\alias{remove_duplicate_subrecords}
\title{Remove duplicate subrecords from a tidyged record}
\usage{
remove_duplicate_subrecords(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of the record to act on.}
}
\value{
The same tidyged object with duplicate subrecords removed from the specified record.
}
\description{
This function removes duplicate structures within a single record.
}
