% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_living}
\alias{remove_living}
\title{Remove living individuals from a tidyged object}
\usage{
remove_living(
  tg,
  max_age = 120,
  guess = FALSE,
  remove_record = FALSE,
  add_note = FALSE,
  note_text = "Information on this individual has been removed",
  remove_supp_records = TRUE,
  update_date_changed = TRUE
)
}
\arguments{
\item{tg}{A tidyged object.}

\item{max_age}{The maximum age to assume for a living person.}

\item{guess}{Whether to guess the age of individuals if no death event or date of birth is given and possibly retain them, or be cautious and remove them anyway (the default).}

\item{remove_record}{Whether to remove the Individual records, or retain them as placeholders.}

\item{add_note}{If the Individual record is kept, whether to include a note in the record (given by the note_text parameter).}

\item{note_text}{If add_note = TRUE, the text of the note.}

\item{remove_supp_records}{Whether to also remove supporting records (sources, notes, multimedia). These may contain names and dates so it is probably best to remove them.}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
A tidyged object cleansed of information on living individuals.
}
\description{
Remove living individuals from a tidyged object
}
