% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arrange_records}
\alias{arrange_records}
\title{Arrange all records in a tidyged object}
\usage{
arrange_records(tg, order = "IFMSRN")
}
\arguments{
\item{tg}{A tidyged object.}

\item{order}{A character string indicating the desired order of records. The letters
indicate (I)ndividual, (F)amily Group, (M)ultimedia, (S)ource, (R)epository, (N)ote.}
}
\value{
An arranged tidyged object.
}
\description{
This function groups together all records of a particular type and puts them in a specific order.
This rearrangement makes no functional difference to the file, it just makes it more organised.
}
