% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_explicit_death_subrecords}
\alias{insert_explicit_death_subrecords}
\title{Insert explicit death subrecords}
\usage{
insert_explicit_death_subrecords(
  tg,
  max_age = 120,
  guess = FALSE,
  explan_note = "This death event has been inferred automatically from other facts"
)
}
\arguments{
\item{tg}{A tidyged object.}

\item{max_age}{The maximum age to assume for a living person.}

\item{guess}{If a date of birth cannot be found, whether to guess it from other information.}

\item{explan_note}{Text to include in an explanatory note for any added death subrecords. An empty string will not add a note.}
}
\value{
An updated tidyged object with additional death subrecords.
}
\description{
This function inserts explicit death subrecords for individuals who have a date of birth that
makes them older than a maximum age.
}
