% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gedcoms.R
\name{merge_gedcoms}
\alias{merge_gedcoms}
\title{Merge two tidyged objects}
\usage{
merge_gedcoms(tg1, tg2)
}
\arguments{
\item{tg1}{The first tidyged object to merge.}

\item{tg2}{The second tidyged object to merge.}
}
\value{
A new tidyged object containing the records of both input objects.
It will also have the same header and submitter information as the first input tidyged object.
}
\description{
Merge two tidyged objects
}
