% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_living}
\alias{remove_living}
\title{Remove data for living individuals in a tidyged object}
\usage{
remove_living(
  tg,
  max_age = 120,
  guess = FALSE,
  remove_record = FALSE,
  explan_note = "Information on this individual has been redacted",
  remove_supp_records = TRUE
)
}
\arguments{
\item{tg}{A tidyged object.}

\item{max_age}{The maximum age to assume for a living person (if a date of birth is given).}

\item{guess}{Whether to guess the age of individuals if no death event or date of birth is given and possibly retain them, or be cautious and remove them anyway (the default).}

\item{remove_record}{Whether to remove the Individual records, or retain them as placeholders.}

\item{explan_note}{Text to include in an explanatory note for any redacted records. An empty string will not add a note.}

\item{remove_supp_records}{Whether to also remove supporting records (sources, repositories, notes, multimedia) associated with the living individuals. These may contain names and dates so it is probably best to remove them.}
}
\value{
A tidyged object cleansed of information on living individuals.
}
\description{
Remove data for living individuals in a tidyged object
}
